bai20buoi5.cpp

+++ bai20buoi5.cpp
@@ -80,7 +80,7 @@ void BurbleSort(float a[], int n){
 	for (int i=0;i<n-1;i++){
 		for (int j=n-1;j>i;j--){
 			if (a[j]<a[i]){
-				int tmp=a[i];
+				float tmp=a[i];
 				a[i]=a[j];
 				a[j]=tmp;
 			}
